find product by id in the domain

+++ src/main/java/com/thoughtworks/com/domain/IProductsCatalog.java
@@ -7,5 +7,5 @@ public interface IProductsCatalog {
 
     public List<Product> getProductList();
 
-    public Product find(long id);
+    public Product find(int id);
 }
